+++ 3rd-party/letter-shell-3.1.2/src/shell_ext.h
@@ -27,7 +27,7 @@ typedef enum
     NUM_TYPE_FLOAT                                          /**< 浮点型 */
 } ShellNumType;
 
-unsigned int shellExtParsePara(Shell *shell, char *string);
+unsigned long shellExtParsePara(Shell *shell, char *string);
 int shellExtRun(Shell *shell, ShellCommand *command, int argc, char *argv[]);
 
 #endif
